+++ CHANGELOG.md
@@ -3,3 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [Unreleased]
+### Added
+ - Basic toolbar controls
+ - Two column layout
+ - Live html rendered preview column
+ - CodeMirror editor
+[Unreleased]: dev..master